Math.MetricSpace

Documentation

class MetricSpace a where

A metric space is a set together with a notion of distance between elements. Distance is computed by a function dist which has the following four laws:

non-negative: forall x y. dist x y >= 0

identity of indiscernibles: forall x y. dist x y == 0 = x == y

symmetry: forall x y. dist x y == dist y x

triangle inequality: forall x y z. dist x z <= dist x y + dist y z

Minimal complete definition

dist

Methods

dist :: Floating b => a -> a -> b

(<->) :: Floating b => a -> a -> b infix 0

Infix notation for dist.

newtype WrappedIntegral a

Constructors

WrappedIntegral

Fields

getA :: Integral a => a