Math.MetricSpace

Documentation

class MetricSpace a where

A metric space is a set together with a notion of distance between elements. Distance is computed by a function dist which has the following four laws:

non-negative: forall x y. dist x y >= 0

identity of indiscernibles: forall x y. dist x y == 0 <=> x == y

symmetry: forall x y. dist x y == dist y x

triangle inequality: forall x y z. dist x z <= dist x y + dist1 y z

Minimal complete definition

dist

Methods

dist :: Floating b => a -> a -> b

(<->) :: Floating b => a -> a -> b infix 0

Infix notation for dist.

newtype Euclidean1 a

Constructors

Euclidean1

Fields

getEuclidean1 :: Integral a => a

newtype Euclidean2 a b

Constructors

Euclidean2

Fields

getEuclidean2 :: (Integral a, Integral b) => (a, b)

newtype Euclidean3 a b c

Constructors

Euclidean3

Fields

getEuclidean3 :: (Integral a, Integral b, Integral c) => (a, b, c)

newtype Euclidean4 a b c d

Constructors

Euclidean4

Fields

getEuclidean4 :: (Integral a, Integral b, Integral c, Integral d) => (a, b, c, d)

newtype Euclidean5 a b c d e

Constructors

Euclidean5

Fields

getEuclidean5 :: (Integral a, Integral b, Integral c, Integral d, Integral e) => (a, b, c, d, e)

newtype Discrete1 a

Constructors

Discrete1

Fields

getDiscrete1 :: a

newtype Discrete2 a b

Constructors

Discrete2

Fields

getDiscrete2 :: (a, b)

newtype Discrete3 a b c

Constructors

Discrete3

Fields

getDiscrete3 :: (a, b, c)

newtype Discrete4 a b c d

Constructors

Discrete4

Fields

getDiscrete4 :: (a, b, c, d)

newtype Discrete5 a b c d e

Constructors

Discrete5

Fields

getDiscrete5 :: (a, b, c, d, e)

newtype Levenshtein

Constructors

Levenshtein String