Math.MetricSpace

Documentation

newtype MetricSpace a b

A metric space is a set together with a notion of distance between elements. Distance is computed by a function dist which has the following four laws:

non-negative: forall x y. dist x y >= 0

identity of indiscernibles: forall x y. dist x y == 0 <=> x == y

symmetry: forall x y. dist x y == dist y x

triangle inequality: forall x y z. dist x z <= dist x y + dist y z

Constructors

MetricSpace

Fields

dist :: a -> a -> b

(<->) :: MetricSpace a b -> a -> a -> b infixl 8

levenshtein :: Floating b => MetricSpace String b

Levenshtein distance between Strings.

discrete :: (Eq a, Floating b) => MetricSpace (Vector a) b

Discrete distance over n-dimensional Vectors.

euclidean :: RealFloat a => MetricSpace (Vector a) a

Euclidean distance over n-dimensional Vectors.

taxicab :: RealFloat a => MetricSpace (Vector a) a

Taxicab distance over n-dimensional Vectors.